+++ src/components/CheckList/PListAddModal.jsx
@@ -7,13 +7,12 @@ import Button from "../Button";
 import {
     toCleanStateContext,
     toCleanDispatchContext,
-} from "../../pages/GroupSpacePage";
+} from "../../context/GroupContext";
 import DatePicker from "react-datepicker";
 
 import { ko } from "date-fns/locale";
 import { registerLocale } from "react-datepicker";
 import DropDown from "./DropDown";
-import { Dropdown, ButtonGroup, DropdownButton } from "react-bootstrap";
 
 registerLocale("ko", ko);
