+++ src/components/groupHome/GroupHome.jsx
@@ -4,8 +4,11 @@ import home_img from "../../assets/home_img.PNG";
 import pencil_img from "../../assets/pencil_img.PNG";
 import Button from "../Button";
 import GEvalItem from "./GEvalItem";
+import { useContext } from "react";
+import { toCleanStateContext } from "../../context/GroupContext";
 
 const GroupHome = () => {
+    const { personData } = useContext(toCleanStateContext);
     return (
         <div className="GroupHome">
             <div className="groupName">
@@ -23,7 +26,9 @@ const GroupHome = () => {
                     <div className="groupEval">
                         <h3>그룹원</h3>
                         <div>
-                            <GEvalItem />;
+                            {personData.map((item) => {
+                                <GEvalItem />;
+                            })}
                         </div>
                         <Button text={"그룹원 평가"} type={"eval"} />
                     </div>
