+++ src/context/GroupContext.js
@@ -1,0 +1,5 @@
+// context/GroupContext.js
+import { createContext } from "react";
+
+export const toCleanStateContext = createContext();
+export const toCleanDispatchContext = createContext();
