+++ src/pages/GroupHomePage.jsx
@@ -1,12 +1,13 @@
 import Header from "../components/Header";
 import GroupHome from "../components/GroupHome/GroupHome";
+import GroupProvider from "../context/GroupProvider";
 
 const GroupHomePage = () => {
     return (
-        <>
+        <GroupProvider>
             <Header />
             <GroupHome />
-        </>
+        </GroupProvider>
     );
 };
